requirements/win.txt
# Production and Development
Django==1.8.4
django-mptt
-e git+https://github.com/zyga/django-pagination.git@master#egg=linaro-django-pagination
django-filter
webstack-django-sorting
djangorestframework
Pillow
django-haystack
pyelasticsearch
django-froala-editor
# -e git+https://github.com/helber/django-dbsettings.git@master#egg=django-dbsettings
django-modeltranslation
-e git+https://github.com/awecode/njango.git@master#egg=njango
sorl-thumbnail
django-redis
jsonfield
WeasyPrint
openpyxl
requests
django-admin-sortable2
django-solo
